pyproject.toml
[project]
name = "finassist"
version = "0.1.0"
description = ""
authors = [
    {name = "MaxGalindo150",email = "maximilianogalindo7@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "google-adk (>=1.1.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
